Support reuse of GenEvent in to_hepmc3 interface (#49)
The optional `to_hepmc3` interface can be added to foreign event objects
to support writing them as Hepmc files. This adds support for reuse of
an existing GenEvent that was already converted. Clients can use this to
add a GenRunInfo only once to the first event, which contains info that
does not change during the run.