Fixes last event not read issue (#64)
Reported by Pim Herbschleb via email.

The HepMC3 C++ has a bug, because under some circumstances it reports
reading an event as success (with read_event()), even if the input
stream did not contain any event data. I am using a workaround for this,
which unfortunately require the use of heuristics that can fail. Pim
reported an input file on which the previous heuristic failed, so I
changed the code once again to pass this and the previous tests. Some
tests needed to be modified. The new code is not able to distinguish
between an EOF error at the end of an event and a reading error in the
middle of an event anymore, but it is more important to read events
correctly which have been correctly written than handling a rare error
cases more elegantly.